requirements.txt
python-binance==1.0.16
typer==0.9.0